+++ src/Logger.ts
@@ -1,6 +1,6 @@
-import { OutputChannel, workspace, window } from 'vscode';
+import * as vscode from 'vscode';
 
-const logChannel: OutputChannel = window.createOutputChannel('Verilog');
+const outputChannel: vscode.OutputChannel = vscode.window.createOutputChannel('Verilog');
 
 export enum LogSeverity {
     info,
@@ -10,30 +10,15 @@ export enum LogSeverity {
 }
 
 export class Logger {
-    isEnabled: boolean;
-
     constructor() {
-        // Register for any changes to logging
-        workspace.onDidChangeConfiguration(() => {
-            this.checkIfEnabled();
-        });
-        this.checkIfEnabled();
-    }
-
-    checkIfEnabled() {
-        this.isEnabled = <boolean>(
-            workspace.getConfiguration().get('verilog.logging.enabled')
-        );
     }
 
     log(msg: string, severity: LogSeverity = LogSeverity.info) {
-        if (this.isEnabled) {
-            if (severity == LogSeverity.command)
-                {logChannel.appendLine('> ' + msg);}
-            else
-                {logChannel.appendLine(
-                    '[' + LogSeverity[severity] + '] ' + msg
-                );}
+        if (severity === LogSeverity.command) { outputChannel.appendLine('> ' + msg); }
+        else {
+            outputChannel.appendLine(
+                '[' + LogSeverity[severity] + '] ' + msg
+            );
         }
     }
 }